Created useUsersPage hook
